[ATS-188] Handle connectionState changes.
